refactor(control): split ws_handler tests to tests.rs (size gate)
